requirements.txt
speedtest-cli
pytest
click

